tighten up declarations in order for this to compile on gcc-3.4 on amd64.

Patches applied:

 * srivasta@debian.org--2003-primary/mailagent--debian--3.0--patch-3
   tighten up declarations in order for this to compile on gcc-3.4 on amd64.


git-archimport-id: srivasta@debian.org--2003-primary/mailagent--devo--3.0--patch-2

+++ agent/pl/utmp/utmp_ph.c
@@ -77,8 +77,8 @@ main()
 	char pack[MAX_LEN];
 	char fields[MAX_LEN];
 	char buf[MAX_LEN];
-	int user_off = (int) utmp->ut_name;		/* Offset of ut_name[] */
-	int line_off = (int) utmp->ut_line;		/* Offset of ut_line[] */
+	size_t user_off = (size_t) utmp->ut_name;		/* Offset of ut_name[] */
+	size_t line_off = (size_t) utmp->ut_line;		/* Offset of ut_line[] */
 	int user_len = sizeof(utmp->ut_name);	/* Length of ut_name[] array */
 	int line_len = sizeof(utmp->ut_line);	/* Length of ut_line[] array */
 	int last_off = 0;						/* Last offset in pack format */
